Moved pass payload to a dedicated namespace
