+++ samples/Smiosoft.PASS.Examples.AspNetCore/Payloads/RabbitMqExampleQueuePayload.cs
@@ -1,6 +1,8 @@
+using Smiosoft.PASS.Payload;
+
 namespace Smiosoft.PASS.Examples.AspNetCore.Payloads
 {
-	public partial class ExampleQueuePayload : IPayload
+	public partial class RabbitMqExampleQueuePayload : IPayload
 	{
 		public string Message { get; set; } = string.Empty;
 	}
