+++ src/Smiosoft.PASS/Publisher/Handler/HandlerWrapper.cs
@@ -1,5 +1,6 @@
 using System.Threading;
 using System.Threading.Tasks;
+using Smiosoft.PASS.Payload;
 
 namespace Smiosoft.PASS.Publisher.Handler
 {
